pyproject.toml
[project]
name = "ai-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.1",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.13",
    "langchain-text-splitters>=1.1.2",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.44.0",
]
